requirements.txt
openai
python-dotenv
hydra-core
omegaconf
numpy
scipy
torch